Move test 078 to general
The cmap table contains no glyph names, fonttools constructs those from
information in the post and CFF tables.
