+++ Lib/fontbakery/specifications/general.py
@@ -622,3 +622,20 @@ def com_google_fonts_check_060(ttFont):
   else:
     yield FAIL, ("The following glyph IDs"
                  " are incorrectly named: {}").format(bad_glyphIDs)
+
+
+@check(id='com.google.fonts/check/078')
+def com_google_fonts_check_078(ttFont):
+  """Check that glyph names do not exceed max length."""
+  if ttFont.sfntVersion == b'\x00\x01\x00\x00' and ttFont.get(
+      "post") and ttFont["post"].formatType == 3.0:
+    yield PASS, ("TrueType fonts with a format 3.0 post table contain no "
+                 "glyph names.")
+  else:
+    failed = False
+    for name in ttFont.getGlyphOrder():
+      if len(name) > 109:
+        failed = True
+        yield FAIL, ("Glyph name is too long:" " '{}'").format(name)
+    if not failed:
+      yield PASS, "No glyph names exceed max allowed length."
